Clean up library directory structure
