+++ .gitignore
@@ -4,6 +4,7 @@
 .idea
 .iml
 *.sublime-workspace
+.DS_Store
 
 # Golang
 vendor
