--- libraries/golang/layering/handler/handler.go
@@ -1,20 +1,0 @@
-package handler
-
-import "web.microservice.shell/libraries/golang/layering/controller"
-
-type BaseHandler struct {
-	controller *controller.BaseController
-}
-
-func (h *BaseHandler) PublicAPIController() controller.PublicAPIController {
-	return h.controller
-}
-
-func (h *BaseHandler) Controller() *controller.BaseController {
-	return h.controller
-}
-
-func (c BaseHandler) PublicAPIClose() {
-	c.controller.PsqlClose()
-	c.controller.SgClose()
-}
